Update UI login dan tambah fitur reset password

+++ database/migrations/2026_04_28_142117_create_katalogs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->json('foto');// Path foto disimpan sebagai string
             $table->integer('berat'); // Integer
             $table->integer('stok')->nullable(); // Sesuai desain UI Anda
-            $table->integer('harga'); // Integer
+            $table->bigInteger('harga'); // Integer
             $table->timestamps();
         });
     }
